qwazr-utils/src/main/java/com/qwazr/utils/concurrent/ReferenceCounter.java
/*
 * Copyright 2015-2017 Emmanuel Keller / QWAZR
 * <p>
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * <p>
 * http://www.apache.org/licenses/LICENSE-2.0
 * <p>
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
package com.qwazr.utils.concurrent;

public interface ReferenceCounter {

	ReferenceCounter acquire();

	int release();

	class Impl implements ReferenceCounter {

		private int counter = 0;

		public synchronized Impl acquire() {
			++counter;
			return this;
		}

		public synchronized int release() {
			return --counter;
		}

	}

}
